--- a/src/math/round.rs
+++ b/src/math/round.rs
@@ -1,38 +1,10 @@
+use super::copysign;
+use super::trunc;
 use core::f64;
 
-const TOINT: f64 = 1.0 / f64::EPSILON;
-
 #[cfg_attr(all(test, assert_no_panic), no_panic::no_panic)]
-pub fn round(mut x: f64) -> f64 {
-    let i = x.to_bits();
-    let e: u64 = i >> 52 & 0x7ff;
-    let mut y: f64;
-
-    if e >= 0x3ff + 52 {
-        return x;
-    }
-    if e < 0x3ff - 1 {
-        // raise inexact if x!=0
-        force_eval!(x + TOINT);
-        return 0.0 * x;
-    }
-    if i >> 63 != 0 {
-        x = -x;
-    }
-    y = x + TOINT - TOINT - x;
-    if y > 0.5 {
-        y = y + x - 1.0;
-    } else if y <= -0.5 {
-        y = y + x + 1.0;
-    } else {
-        y = y + x;
-    }
-
-    if i >> 63 != 0 {
-        -y
-    } else {
-        y
-    }
+pub fn round(x: f64) -> f64 {
+    trunc(x + copysign(0.5 - 0.25 * f64::EPSILON, x))
 }
 
 #[cfg(test)]
@@ -42,5 +14,15 @@ mod tests {
     #[test]
     fn negative_zero() {
         assert_eq!(round(-0.0_f64).to_bits(), (-0.0_f64).to_bits());
+    }
+
+    #[test]
+    fn sanity_check() {
+        assert_eq!(round(-1.0), -1.0);
+        assert_eq!(round(2.8), 3.0);
+        assert_eq!(round(-0.5), -1.0);
+        assert_eq!(round(0.5), 1.0);
+        assert_eq!(round(-1.5), -2.0);
+        assert_eq!(round(1.5), 2.0);
     }
 }