In reference to #154, this optimizes the `round` and `roundf` functions.  This does not change the behavior at all.

I compared all f32 values to verify that the behavior of `roundf` has not changed.

For `round`, I directly translated the `roundf` function and verified that it behaved the same with random bits for the sign and exponent and beginning and end of the mantissa.

I ran some criterion benchmarks and basically both functions went from about 7ns to about 3ns.

```
round/round             time:   [7.1414 ns 7.1470 ns 7.1543 ns]                         
round/round_new         time:   [3.0178 ns 3.0311 ns 3.0471 ns]                             

roundf/roundf           time:   [6.8636 ns 6.8767 ns 6.9016 ns]                           
roundf/roundf_new       time:   [2.8326 ns 2.8401 ns 2.8562 ns]                               
```