@Lokathor 2021-07-01
Looks like we had some CI trouble, but actually we haven't been passing CI lately to begin with.

It appears that none of the failures were because of thia PR.

My only thought is that in general we might want to mark these and other functions as `inline`, but that can be done later.

@bvinc I can merge if you're all done.