@bvinc 2021-07-02
I'm done, but I'll add #inline too if you want.  Looks like it should definitely be on copysign/copysignf.  Are you sure we also want inline on trunc/truncf?