@Lokathor 2021-07-02
oh i forgot how big that one is.

we'll do inline separate, I'll just merge for now.